First Commit
Added a 3-tier architecture to the Solution where all contexts must be
opened in the DataLayer. The DataLayer is then consumed by the Business
Layer and the latter is consumed by both the App and API if implemented.

The Edmx was created in the Data Layer but due to the fact that the
models must be accessable in all class libraries, the DbContext tt model
classes where moved to the Common project while maintaining the
important parts in the DataLayer. Moreover, the EntityContext was set to
internal.

The App is the default Visual Studio template but I added the
RoleManager.